requirements.txt
google-cloud-bigquery==3.9.0
google-cloud-storage==2.7.0